Breaking out the canvas related logic onto a worker thread. All trait selection is done by the main process, and then dealt out to child processes to stack the layers. This uncaps the resource usage, meaning that collections can be generated significantly faster.
