+++ package.json
@@ -38,6 +38,7 @@
     "eslint-plugin-node": "^11.1.0",
     "keccak256": "^1.0.3",
     "sharp": "^0.30.2",
-    "gif-encoder-2": "^1.0.5"
+    "gif-encoder-2": "^1.0.5",
+    "workerpool": "^6.3.1"
   }
 }
